archives/lightstream/pyproject.toml
[project]
name = "lightstream"
version = "0.1.0"
description = "PoW DDos protection"
requires-python = ">=3.13.2"
dependencies = [
  "argon2-cffi==23.1.0",
  "cryptography==45.0.2",
  "fastapi[standard]==0.115.12",
  "httpx==0.28.1",
  "pyjwt==2.10.1",
  "pytest==8.3.5",
]
